# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampaignCell API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with GSM + OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /v1/auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP for GSM number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequestInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResponse"
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and revoke refresh token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── CAMPAIGNS ──────────────────────────────────────────────────────────────
  /v1/campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: segment
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignListResponse"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /v1/campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/campaigns/{id}/analyze:
    post:
      operationId: analyzeCampaign
      tags: [campaigns]
      summary: Trigger AI analysis for campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
  /v1/campaigns/{id}/publish:
    post:
      operationId: publishCampaign
      tags: [campaigns]
      summary: Publish a campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /v1/campaigns/{id}/archive:
    post:
      operationId: archiveCampaign
      tags: [campaigns]
      summary: Archive a campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ─── OPTIMIZATION CASES ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # ── AUTH SCHEMAS ──────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [gsmNumber, otp]
      properties:
        gsmNumber:
          type: string
        otp:
          type: string
    OtpRequestInput:
      type: object
      required: [gsmNumber]
      properties:
        gsmNumber:
          type: string
    OtpRequestResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, name, email, role, gsmNumber]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [SUBSCRIBER, CAMPAIGN_EXPERT, SUPERVISOR, ADMIN]
        gsmNumber:
          type: string
        isLocked:
          type: boolean
        createdAt:
          type: string

    # ── USER SCHEMAS ──────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      required: [id, name, type, status, segment, priority, discount, startDate, endDate, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [EK_PAKET, TARIFE_YUKSELTME, CIHAZ_FIRSATI, SADAKAT]
        status:
          type: string
          enum: [DRAFT, ACTIVE, MANUAL_OPTIMIZATION_REQUIRED, OPTIMIZING, AB_TESTING, OPTIMIZED, PUBLISHED, ARCHIVED]
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
        priority:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        discount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        recommendationScore:
          type: ["number", "null"]
        conversionProbability:
          type: ["number", "null"]
        aiReasoning:
          type: ["string", "null"]
        isAiAnalyzed:
          type: boolean
        conversionRate:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CampaignInput:
      type: object
      required: [name, type, discount, startDate, endDate]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [EK_PAKET, TARIFE_YUKSELTME, CIHAZ_FIRSATI, SADAKAT]
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
        discount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
        priority:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        discount:
          type: number
    CampaignListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AiAnalysisResult:
      type: object
      required: [success, aiAvailable]
      properties:
        success:
          type: boolean
        aiAvailable:
          type: boolean
        message:
          type: ["string", "null"]
        campaign:
          $ref: "#/components/schemas/Campaign"

    # ── OPTIMIZATION CASE SCHEMAS ─────────────────────────────────────────────
...